Namespace and split up.
