+++ src/gui/dataTableModelInterface.cpp
@@ -4,9 +4,9 @@
 #include "gui/dataTableModelInterface.h"
 #include "procedure/nodes/node.h"
 
-DataTableModelInterface::DataTableModelInterface(DataModelBase &dataModel) : dataModel_(dataModel)
+DataTableModelInterface::DataTableModelInterface(DataModel::Base &dataModel) : dataModel_(dataModel)
 {
-    dataModel_.setMutationSignalFunction([this](DataModelBase::MutationSignal signal, int startIndex, int endIndex)
+    dataModel_.setMutationSignalFunction([this](DataModel::Base::MutationSignal signal, int startIndex, int endIndex)
                                          { dataMutated(signal, startIndex, endIndex); });
 }
 
@@ -25,9 +25,9 @@ Qt::ItemFlags DataTableModelInterface::flags(const QModelIndex &index) const
 {
     auto &propertyFlags = dataModel_.propertyFlags(index.column());
     Qt::ItemFlags flags = Qt::ItemIsSelectable;
-    if (!propertyFlags.isSet(DataItemProperty::ReadOnly))
+    if (!propertyFlags.isSet(DataModel::ItemProperty::ReadOnly))
         flags.setFlag(Qt::ItemIsEditable);
-    if (!propertyFlags.isSet(DataItemProperty::Disabled))
+    if (!propertyFlags.isSet(DataModel::ItemProperty::Disabled))
         flags.setFlag(Qt::ItemIsEnabled);
     return flags;
 }
@@ -51,31 +51,32 @@ QVariant DataTableModelInterface::data(const QModelIndex &index, int role) const
     auto property = dataModel_.getProperty(index.row(), index.column());
 
     // Construct a QVariant from the contents of our std::variant
-    return std::visit(DataItemVisitor{[](int arg) { return QVariant(arg); }, [](double arg) { return QVariant(arg); },
-                                      [](std::string_view arg) { return QVariant(QString::fromStdString(std::string(arg))); },
-                                      [](std::string &arg) { return QVariant(QString::fromStdString(arg)); }},
-                      property);
+    return std::visit(
+        DataModel::PropertyVisitor{[](int arg) { return QVariant(arg); }, [](double arg) { return QVariant(arg); },
+                                   [](std::string_view arg) { return QVariant(QString::fromStdString(std::string(arg))); },
+                                   [](std::string &arg) { return QVariant(QString::fromStdString(arg)); }},
+        property);
 }
 
 // Set data for the index and role specified
 bool DataTableModelInterface::setData(const QModelIndex &index, const QVariant &value, int role)
 {
-    if (role != Qt::EditRole || dataModel_.isPropertyFlagSet(index.column(), DataItemProperty::PropertyFlag::ReadOnly))
+    if (role != Qt::EditRole || dataModel_.isPropertyFlagSet(index.column(), DataModel::ItemProperty::PropertyFlag::ReadOnly))
         return false;
 
     // Set new value
     bool success = false;
     switch (dataModel_.propertyType(index.column()))
     {
-        case (DataItemProperty::PropertyType::Integer):
-            success = dataModel_.setProperty(index.row(), index.column(), DataModelBase::PropertyValue(value.toInt()));
+        case (DataModel::ItemProperty::PropertyType::Integer):
+            success = dataModel_.setProperty(index.row(), index.column(), DataModel::PropertyValue(value.toInt()));
             break;
-        case (DataItemProperty::PropertyType::Double):
-            success = dataModel_.setProperty(index.row(), index.column(), DataModelBase::PropertyValue(value.toDouble()));
+        case (DataModel::ItemProperty::PropertyType::Double):
+            success = dataModel_.setProperty(index.row(), index.column(), DataModel::PropertyValue(value.toDouble()));
             break;
-        case (DataItemProperty::PropertyType::String):
-            success = dataModel_.setProperty(index.row(), index.column(),
-                                             DataModelBase::PropertyValue(value.toString().toStdString()));
+        case (DataModel::ItemProperty::PropertyType::String):
+            success =
+                dataModel_.setProperty(index.row(), index.column(), DataModel::PropertyValue(value.toString().toStdString()));
             break;
         default:
             Messenger::error("DataTableModelInterface doesn't know how to handle this PropertyType.\n");
@@ -123,20 +124,20 @@ bool DataTableModelInterface::removeRows(int row, int count, const QModelIndex &
  */
 
 // React to a mutation in the model
-void DataTableModelInterface::dataMutated(DataModelBase::MutationSignal signal, int startIndex, int endIndex)
+void DataTableModelInterface::dataMutated(DataModel::Base::MutationSignal signal, int startIndex, int endIndex)
 {
     switch (signal)
     {
-        case (DataModelBase::MutationSignal::DataCreationStarted):
+        case (DataModel::Base::MutationSignal::DataCreationStarted):
             beginInsertRows({}, startIndex, endIndex);
             break;
-        case (DataModelBase::MutationSignal::DataCreationFinished):
+        case (DataModel::Base::MutationSignal::DataCreationFinished):
             endInsertRows();
             break;
-        case (DataModelBase::MutationSignal::DataRemovalStarted):
+        case (DataModel::Base::MutationSignal::DataRemovalStarted):
             beginRemoveRows({}, startIndex, endIndex);
             break;
-        case (DataModelBase::MutationSignal::DataRemovalFinished):
+        case (DataModel::Base::MutationSignal::DataRemovalFinished):
             endRemoveRows();
             break;
     }
